fix error on PL SQL input parameter
The data always truncated.   After hours investigating,  that one is the PL SQL input parameter declaration.  IN declared as VARCHAR (200),  meanwhile I focus on the table, even make it to VARCHAR (1000),  data is truncated at the Input param, not at table..
Experience will be the best teacher..  next time if got similar problem